feat: 보드 일급 컬렉션 기능 구현
- 팀의 왕 상태 기능 구현
- 보드 상태 기능 구현
- 팀별 기물 점수 기능 구현
- 보드 현재 기물 배치 조회 기능 구현
